requirements.txt
asgiref==3.7.2
certifi==2023.11.17
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2023.3.post1
setuptools==69.0.2
sqlparse==0.4.4
wheel==0.42.0
